Refactor packaging for echo nukleus specification
