+++ echo/src/test/java/org/kaazing/nuklei/echo/internal/control/ControlIT.java
@@ -36,7 +36,7 @@ import uk.co.real_logic.agrona.concurrent.ringbuffer.RingBufferDescriptor;
 public class ControlIT
 {
     private final K3poRule k3po = new K3poRule()
-        .setScriptRoot("org/kaazing/nuklei/specification/echo/control");
+        .setScriptRoot("org/kaazing/specification/nuklei/echo/control");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
