+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/control/bind.source/nukleus.rpt
