+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/control/connect.destination/controller.rpt
