+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/control/prepare.destination/nukleus.rpt
