+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/accept/echo.source.data/source.rpt
