+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/accept/establish.connection/source.rpt
