+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/accept/initiate.nukleus.close/source.rpt
