+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/accept/initiate.source.close/source.rpt
