+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/connect/echo.destination.data/destination.rpt
