+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/connect/establish.connection/destination.rpt
