+++ specification/echo/src/main/scripts/org/kaazing/specification/nuklei/echo/streams/connect/initiate.nukleus.close/nukleus.rpt
