+++ specification/echo/src/test/java/org/kaazing/specification/nuklei/echo/control/ControlIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kaazing.nuklei.specification.echo.control;
+package org.kaazing.specification.nuklei.echo.control;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
